started moved freedv_700 functions out of freedv_api.c
